renamed and some cleanup
