+++ src/Blt/Plugin/Commands/PhpUnitCommands.php
@@ -12,139 +12,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  *
  * @package Example\Blt\Plugin\Commands
  */
-class SwsTestCommands extends BltTasks {
-
-  /**
-   * Runs all tests, including Behat, PHPUnit, and security updates check.
-   *
-   * @command tests
-   *
-   * @aliases ta test tests:all
-   */
-  public function tests() {
-    $this->invokeCommands([
-      'tests:behat:run',
-      'tests:phpunit:coverage:run',
-      'tests:codeception:run',
-      'tests:drupal:run',
-      'tests:security:check:updates',
-      'tests:security:check:composer',
-      'tests:frontend:run',
-    ]);
-  }
-
-  /**
-   * Run all the codeception tests defined in blt.yml.
-   *
-   * @param array $options
-   *   Keyed array of command options.
-   *
-   * @return \Robo\Result
-   *   Result of the test.
-   *
-   * @command tests:codeception:run
-   * @aliases tests:codeception codeception
-   * @options test The key of the tests to run.
-   * @options suite only run a specific suite instead of all suites.
-   */
-  public function runCodeceptionTests(array $options = [
-    'test' => NULL,
-    'suite' => NULL,
-    'group' => NULL,
-  ]) {
-    $failed_test = NULL;
-
-    $tests = $this->getConfigValue('tests.codeception', []);
-
-    // Run only the test that was defined in the options.
-    if (!empty($options['test'])) {
-      $tests = [$options['test'] => $tests[$options['test']]];
-    }
-
-    foreach ($tests as $test) {
-
-      // Filter out the suites that the options doesn't want.
-      $test['suites'] = array_filter($test['suites'], function ($suite) use ($options) {
-        return empty($options['suite']) || strpos($options['suite'], $suite) !== FALSE;
-      });
-
-      // Run each suite of tests.
-      foreach ($test['suites'] as $suite) {
-        $this->say("Running <comment>$suite</comment> Tests.");
-        $test_result = $this->runCodeceptionTestSuite($suite, $test['directory']);
-
-        if (!$test_result->wasSuccessful()) {
-          $failed_test = $test_result;
-        }
-      }
-    }
-    return $failed_test ?: $test_result;
-  }
-
-  /**
-   * Execute codeception test suite.
-   *
-   * @param string $suite
-   *   Codeception suite to run.
-   * @param string $test_directory
-   *   Directory to codeception tests.
-   *
-   * @return \Robo\Result|\Robo\ResultData
-   *   Result of the test.
-   */
-  protected function runCodeceptionTestSuite($suite, $test_directory) {
-    if (!file_exists("$test_directory/$suite/")) {
-      return new ResultData(ResultData::EXITCODE_OK, 'No tests to execute for suite ' . $suite);
-    }
-
-    $root = $this->getConfigValue('repo.root');
-    if (!file_exists("$root/tests/codeception.yml")) {
-      $this->taskFilesystemStack()
-        ->copy("$root/tests/codeception.dist.yml", "$root/tests/codeception.yml")
-        ->run();
-      $this->getConfig()
-        ->expandFileProperties("$root/tests/codeception.yml");
-    }
-
-    $new_test_dir = "$root/tests/codeception/$suite/" . date('Ymd-Hi');
-    $tasks[] = $this->taskFilesystemStack()
-      ->symlink("$test_directory/$suite/", $new_test_dir);
-
-    $test = $this->taskExec('vendor/bin/codecept')
-      ->arg('run')
-      ->arg($suite)
-      ->option('steps')
-      ->option('config', 'tests', '=')
-      ->option('fail-fast')
-      ->option('override', "paths: output: ../artifacts/$suite", '=')
-      ->option('html')
-      ->option('xml');
-
-    if ($group = $this->input()->getOption('group')) {
-      $test->option('group', $group, '=');
-    }
-
-    if ($this->input()->getOption('verbose')) {
-      $test->option('debug');
-      $test->option('verbose');
-    }
-    $tasks[] = $test;
-    $test_result = $this->collectionBuilder()->addTaskList($tasks)->run();
-    // Regardless if the test failed or succeeded, always clean up the temporary
-    // test directory.
-    $this->taskDeleteDir($new_test_dir)->run();
-
-    // Delete the failed file because codeception will try to look for the file
-    // that failed again on the next run. Since we have temporary test
-    // directories we don't want to save that data.
-    foreach (glob("$root/artifacts/*/failed") as $file) {
-      $this->taskFilesystemStack()
-        ->remove($file)
-        ->run();
-    }
-
-    return $test_result;
-  }
+class PhpUnitCommands extends BltTasks {
 
   /**
    * Setup and run PHPUnit tests with code coverage.
